possible_requirements.txt
backcall==0.1.0
beautifulsoup4==4.6.3
certifi==2018.10.15
chardet==3.0.4
decorator==4.3.0
idna==2.7
ipython==7.0.1
ipython-genutils==0.2.0
jedi==0.13.1
mysql-connector-python==8.0.12
numpy==1.15.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==2.0.6
protobuf==3.6.1
ptyprocess==0.6.0
Pygments==2.2.0
requests==2.19.1
scikit-learn==0.20.0
scipy==1.1.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
traitlets==4.3.2
urllib3==1.23
wcwidth==0.1.7
